fix: two CI-blocking memory and float-precision bugs
QoreIRInterpreter: clear slot bits after discard in discardContainerValueSlot
=============================================================================
The container-slot cleanup helper used by HashKeyStore / ListIndexStore had
three branches:
  - container-owned: removeCleanupEntry + clear bits
  - owns_slot_ref + removable cleanup entry: discard ONLY (no bits clear)
  - everything else: removeCleanupEntry + clear bits

The middle branch dropped the slot's +1 reference but left the now-dangling
pointer in values[vid].  Any later instruction that reads that slot — most
visibly the Return handler at line 11630 doing
`values[ret->value.id].discard(xsink)` — dereferenced freed memory.

Manifestation: Swagger.qtest crashed at startup under both --exec-mode=ir
AND --exec-mode=jit on Linux x86_64 (passed on macOS ARM64 because the
heap allocator there didn't immediately reuse the freed node).  Stack:

  #0 main_arena (freed-memory deref)
  #1 AbstractQoreNode::deref
  #2 QoreValue::discard
  #3 QoreIRInterpreter::execute  ← Return handler

Always clearing the slot after discard restores the standard pattern used
by the other branches and every other discard site in the file.

QoreLogicalComparisonOperatorNode: 64-bit -> 32-bit float truncation
====================================================================
`getAsFloat()` returns a 64-bit double (~15 decimal digits of precision),
but doComparison was storing the result into a `float` (32-bit single
precision, ~7 digits).  Near saturation, distinct doubles such as
0.999999998 and 0.999999996 (delta 2e-9) round to the same single-precision
representation, so the spaceship `<=>` reported them as equal — silently
breaking sort stability for any algorithm that compares close doubles.

Manifestation: QoreTokenizerUtils.qtest's CrossEncoderReranker ONNX tensor
inference assertion "ONNX reranker hits should be sorted" failed on Linux
CI.  The dummy reranker model saturates sigmoid for both passages; sort's
comparator returned 0 for unequal doubles, keeping the original order
which happened to put the lower score first.  Local macOS ARM64 differs
just enough in ONNX float output to dodge the truncation collision.

Fixed by widening to `double` at all three call sites in doComparison()
(number-vs-float, float-vs-number, and float-vs-float).  No other location
in lib/ truncates a getAsFloat() result into single precision.

Verified: full ir/ test directory passes (27/27); JITSmoke 164/164,
AOTSmoke 212/212, TieredSmoke 25/25; Swagger.qtest succeeds in both ir
and jit exec modes; QoreTokenizerUtils.qtest 28/28 (was 27/28).

+++ lib/QoreLogicalComparisonOperatorNode.cpp
@@ -151,7 +151,7 @@ int QoreLogicalComparisonOperatorNode::doComparison(const QoreValue& left, const
                 return ln->equals(*rn) ? 0 : 1;
             }
             case NT_FLOAT: {
-                float f = r.getAsFloat();
+                double f = r.getAsFloat();
                 if (std::isnan(f)) {
                     xsink->raiseException("NAN-COMPARE-ERROR", "NaN in floating-point value on right hand side of logical comparison operator");
                     return 0;
@@ -190,7 +190,7 @@ int QoreLogicalComparisonOperatorNode::doComparison(const QoreValue& left, const
 
         switch (lt) {
             case NT_FLOAT: {
-                float lf = l.getAsFloat();
+                double lf = l.getAsFloat();
                 if (std::isnan(lf)) {
                     xsink->raiseException("NAN-COMPARE-ERROR", "NaN in floating-point value on left hand side of logical comparison operator");
                     return 0;
@@ -219,13 +219,18 @@ int QoreLogicalComparisonOperatorNode::doComparison(const QoreValue& left, const
     }
 
     if (lt == NT_FLOAT || rt == NT_FLOAT) {
-        float lf = l.getAsFloat();
+        // Use double (not float) — getAsFloat() returns a 64-bit double; storing
+        // it in a 32-bit float silently truncates ~9 decimal digits of precision,
+        // so two distinct doubles that round to the same single-precision
+        // representation (e.g. 0.999999998 and 0.999999996 near saturation)
+        // would compare as equal, breaking sort stability.
+        double lf = l.getAsFloat();
         if (std::isnan(lf)) {
             xsink->raiseException("NAN-COMPARE-ERROR", "NaN in floating-point value on left hand side of logical comparison operator");
             return 0;
         }
 
-        float rf = r.getAsFloat();
+        double rf = r.getAsFloat();
         if (std::isnan(rf)) {
             xsink->raiseException("NAN-COMPARE-ERROR", "NaN in floating-point value on right hand side of logical comparison operator");
             return 0;
